decimal: cleanup files
