+++ go/vt/vtgate/evalengine/internal/decimal/decimal.go
@@ -22,13 +22,10 @@ limitations under the License.
 package decimal
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
 	"math"
 	"math/big"
 	"strconv"
-	"strings"
 )
 
 // MyMaxScale is the largest scale on a decimal that MySQL supports
@@ -164,213 +161,6 @@ func NewFromUint(value uint64) Decimal {
 		value: new(big.Int).SetUint64(value),
 		exp:   0,
 	}
-}
-
-var errOverflow = errors.New("overflow")
-
-func parseDecimal64(s []byte) (Decimal, error) {
-	const cutoff = math.MaxUint64/10 + 1
-	var n uint64
-	var dot = -1
-
-	for i, c := range s {
-		var d byte
-		switch {
-		case c == '.':
-			if dot > -1 {
-				return Decimal{}, fmt.Errorf("too many .s")
-			}
-			dot = i
-			continue
-		case '0' <= c && c <= '9':
-			d = c - '0'
-		default:
-			return Decimal{}, fmt.Errorf("unexpected character %q", c)
-		}
-
-		if n >= cutoff {
-			// n*base overflows
-			return Decimal{}, errOverflow
-		}
-		n *= 10
-		n1 := n + uint64(d)
-		if n1 < n {
-			return Decimal{}, errOverflow
-		}
-		n = n1
-	}
-
-	var exp int32
-	if dot != -1 {
-		exp = -int32(len(s) - dot - 1)
-	}
-	return Decimal{
-		value: new(big.Int).SetUint64(n),
-		exp:   exp,
-	}, nil
-}
-
-func NewFromMySQL(s []byte) (Decimal, error) {
-	var original = s
-	var neg bool
-
-	if len(s) > 0 {
-		switch s[0] {
-		case '+':
-			s = s[1:]
-		case '-':
-			neg = true
-			s = s[1:]
-		}
-	}
-
-	if len(s) == 0 {
-		return Decimal{}, fmt.Errorf("can't convert %q to decimal: too short", original)
-	}
-
-	if len(s) <= 18 {
-		dec, err := parseDecimal64(s)
-		if err == nil {
-			if neg {
-				dec.value.Neg(dec.value)
-			}
-			return dec, nil
-		}
-		if err != errOverflow {
-			return Decimal{}, fmt.Errorf("can't convert %s to decimal: %v", original, err)
-		}
-	}
-
-	var fractional, integral []byte
-	if pIndex := bytes.IndexByte(s, '.'); pIndex >= 0 {
-		if bytes.IndexByte(s[pIndex+1:], '.') != -1 {
-			return Decimal{}, fmt.Errorf("can't convert %s to decimal: too many .s", original)
-		}
-		if pIndex+1 < len(s) {
-			integral = s[:pIndex]
-			fractional = s[pIndex+1:]
-		} else {
-			integral = s[:pIndex]
-		}
-	} else {
-		integral = s
-	}
-
-	// Check if the size of this bigint would fit in the limits
-	// that MySQL has by default. To do that, we must convert the
-	// length of our integral and fractional part to "mysql digits"
-	myintg := myBigDigits(int32(len(integral)))
-	myfrac := myBigDigits(int32(len(fractional)))
-	if myintg > MyMaxBigDigits {
-		return largestForm(MyMaxPrecision, 0, neg), nil
-	}
-	if myintg+myfrac > MyMaxBigDigits {
-		fractional = fractional[:int((MyMaxBigDigits-myintg)*9)]
-	}
-	value, err := parseLargeDecimal(integral, fractional)
-	if err != nil {
-		return Decimal{}, err
-	}
-	if neg {
-		value.Neg(value)
-	}
-	return Decimal{value: value, exp: -int32(len(fractional))}, nil
-}
-
-// NewFromString returns a new Decimal from a string representation.
-// Trailing zeroes are not trimmed.
-//
-// Example:
-//
-//     d, err := NewFromString("-123.45")
-//     d2, err := NewFromString(".0001")
-//     d3, err := NewFromString("1.47000")
-//
-func NewFromString(value string) (Decimal, error) {
-	originalInput := value
-	var intString string
-	var exp int64
-
-	// Check if number is using scientific notation
-	eIndex := strings.IndexAny(value, "Ee")
-	if eIndex != -1 {
-		expInt, err := strconv.ParseInt(value[eIndex+1:], 10, 32)
-		if err != nil {
-			if e, ok := err.(*strconv.NumError); ok && e.Err == strconv.ErrRange {
-				return Decimal{}, fmt.Errorf("can't convert %s to decimal: fractional part too long", value)
-			}
-			return Decimal{}, fmt.Errorf("can't convert %s to decimal: exponent is not numeric", value)
-		}
-		value = value[:eIndex]
-		exp = expInt
-	}
-
-	pIndex := -1
-	vLen := len(value)
-	for i := 0; i < vLen; i++ {
-		if value[i] == '.' {
-			if pIndex > -1 {
-				return Decimal{}, fmt.Errorf("can't convert %s to decimal: too many .s", value)
-			}
-			pIndex = i
-		}
-	}
-
-	if pIndex == -1 {
-		// There is no decimal point, we can just parse the original string as
-		// an int
-		intString = value
-	} else {
-		if pIndex+1 < vLen {
-			intString = value[:pIndex] + value[pIndex+1:]
-		} else {
-			intString = value[:pIndex]
-		}
-		expInt := -len(value[pIndex+1:])
-		exp += int64(expInt)
-	}
-
-	var dValue *big.Int
-	// strconv.ParseInt is faster than new(big.Int).SetString so this is just a shortcut for strings we know won't overflow
-	if len(intString) <= 18 {
-		parsed64, err := strconv.ParseInt(intString, 10, 64)
-		if err != nil {
-			return Decimal{}, fmt.Errorf("can't convert %s to decimal", value)
-		}
-		dValue = big.NewInt(parsed64)
-	} else {
-		dValue = new(big.Int)
-		_, ok := dValue.SetString(intString, 10)
-		if !ok {
-			return Decimal{}, fmt.Errorf("can't convert %s to decimal", value)
-		}
-	}
-
-	if exp < math.MinInt32 || exp > math.MaxInt32 {
-		// NOTE(vadim): I doubt a string could realistically be this long
-		return Decimal{}, fmt.Errorf("can't convert %s to decimal: fractional part too long", originalInput)
-	}
-
-	return Decimal{
-		value: dValue,
-		exp:   int32(exp),
-	}, nil
-}
-
-// RequireFromString returns a new Decimal from a string representation
-// or panics if NewFromString would have returned an error.
-//
-// Example:
-//
-//     d := RequireFromString("-123.45")
-//     d2 := RequireFromString(".0001")
-//
-func RequireFromString(value string) Decimal {
-	dec, err := NewFromString(value)
-	if err != nil {
-		panic(err)
-	}
-	return dec
 }
 
 // NewFromFloat converts a float64 to Decimal.
